wado-compiler: extract task_return submodule from cm_binding
Move the `task return` expansion / stripping pass into
`synthesis/cm_binding/task_return.rs`. Functions consumed by the
sibling adapter synthesizers stay accessible as `pub(super)`.
